Generate Merged Location And Info data.table From Telemetry Object/List

A Telemetry list hold mutiple animal data in separate list items, each item have the animal location data in a data frame, and other information in various slots. This structure supports flexible S3 methods for telemetry object. However to plot multiple animals location together with ggplot2 it's better to merge all location data into single data frame with an animal id column.

merge_animals(tele_obj_list)

Arguments

tele_obj_list
telemetry object/list

Value

list of - data: all animals merged in one data.table - info: animal information table

Details

We chose this data structure to be the main structure and made the app to work on a set of animals at the same time in all steps. Thus any input telemetry object/List need to be merged into a data.table of location data, and another information data.table for animals. data.table is chosen over data.frame for much better performance.

Examples

merge_animals(buffalo)

#> Error in wrap_single_telemetry(tele_obj_list): object 'buffalo' not found