Parallel fitting models on telemetry list

par_try_models run ctmm::ctmm.select() on each object of list on parallel.

par_fit_models run ctmm::ctmm.fit() on each object of list on parallel.

par_try_models(tele_list, cores = NULL, parallel = TRUE)

par_fit_models(tele_list, cores = NULL, parallel = TRUE)

Arguments

tele_list
ctmm::as.telemetry() telemetry list

cores
the core count to be used for cluster. Could be a positive integer or

Default NULL value will indicate to use a heuristic value based on detected cores, which is roughly min(input_size, physical_cores_count * n), n being 2 for windows, 4 for Mac/Linux. See parallel::detectCores() for more information on physical/logical cores in different platforms.

A negative value like -2 will use all available cores - 2, so that 2 cores are reserved for user's other tasks.

parallel
Use regular lapply() when FALSE. You may notice more console messages in this mode because the console messages in parallel mode are lost by default as they happened in other threads.

Value

par_try_models: list of items named by animal names, each item hold the attempted models as sub items with model type as name.

par_fit_models: list of models named by animal names.

Functions

par_fit_models: