Parallel Apply Function To List In All Platforms

This is a generic parallel lapply that work across all major platforms.

para_ll(ll, fun, fallback = FALSE, win_init = ctmmweb:::WIN_INIT_ctmm)

Arguments

ll
Input list.

fun
Function to be applied on ll. Note only single parameter function is accepted, otherwise it's difficult to determine how to assign input parameters to each list item and worker. You need to convert multiple parameter function into a function take single list parameter, and assign parameters in that list accordingly. align_list() is a helper function to align two lists.

fallback
Use regular lapply. This is used to test if parallel caused problems.

win_init
Expression to be initialized in Windows. Because all parameters should be included in the input list already, this usually means library calls, like {library(ctmm)} for ctmm related operations, which has been taken care of with the default value ctmmweb:::WIN_INIT_ctmm.

Value

List of applied results

Details

In Windows parallel::parLapplyLB is used, which is a socket cluster and need to initialize each session manually with win_init if needed. In Linux/Mac parallel::mclapply is used, where each worker will inherit the current environment through forking, so no additional setup is required.