Plot a group of empirical or modeled variograms.

This is a wrapper over ctmm::plot.variogram() to make group plot easier.

plot_vario(
  vario_list,
  model_list = NULL,
  title_vec = NULL,
  fraction = 0.5,
  relative_zoom = TRUE,
  cex = 0.65,
  model_color = "blue",
  columns = 2
)

Arguments

vario_list
list of ctmm::variogram(). The names of list may be needed for figure titles.

model_list
CTMM models list. The names of list may be needed for figure titles. Draw modeled variogram if provided. The models list should match vario_list in length and animal, so that ith model is for ith variogram. Possible values:

It can be skipped to use default value NULL, and only empirical variograms are drawn. vario_list names will be used as figure title. Better use named parameters for others in this case.

list of guesstimated models from ctmm::ctmm.guess() on telemetry data, overlay guesstimate variogram.

list of fitted models from ctmm::ctmm.select() on vario_list, overlay modeled variogram.

It can be a list of list, with each item provide list of CTMM objects that to be plotted in each variogram. model_color can be a vector to specify color for each model.

title_vec
vector of figure titles. If not provided, names of vario_list or model_list will be used.

fraction
Fraction of time-lag range, 0 ~ 1.

relative_zoom

Relative mode zoom every plot by fraction of their own Time-lag range. The X, Y axes are not synced.

Absolute mode operate on the max Time-lag range individual in group, and all others scaled with same X, Y axes for easier comparison.

cex
The magnification factor of plot text and symbols. See par(). You may need a smaller value if figure margins is too large

model_color
The color of model variogram. It'll be same for every plot.

columns
The columns of the group plot layout.