Build maps of animal locations

An interactive map will shown in RStudio Viewer pane when running in interactive session. You can also further augment it with leaflet functions, or save to a html with htmlwidgets::saveWidget().

point_map(loc_data_subset)

range_map(hrange_list, hr_levels, hr_color_vec)

point_range_map(loc_data_subset, hrange_list, hr_levels, hr_color_vec)

heat_map(loc_data_subset)

Arguments

loc_data_subset
data.table subset of full data set of animal locations from combine(). The id column need to keep all animal names in factor levels to maintain color consistency.

hrange_list
list of home range UD object. The names of list will be used as layer names, usually they are model names.

hr_levels
The vector of level.UD in ctmm::plot.telemetry. To be consistent with ctmm they are values 0 ~ 1 (for example 0.95). Note the app UI take values 0 ~ 100 (for example 95) for easier input.

hr_color_vec
Vector of color names to be used for each home range. The length of hr_color_vec should match length of hrange_list.

Value

A Leaflet map widget.

Functions

point_map: Build point map of animal locations

range_map: Build map of home ranges

point_range_map: Build map of home ranges and animal locations

heat_map: Build heat map of animal locations