Build home range map with animal locations

An interactive map will shown in RStudio Viewer pane when running in interactive session. You can also further augment it with leaflet functions, or save to a html with htmlwidgets::saveWidget.

point_range_map(dt_subset, hrange_list, hr_levels, hr_color_vec)

Arguments

dt_subset
data.table subset of full data set of animal locations from merge_tele. The id column need to keep all animal names in levels to maintain color consistency.

hrange_list
list of home range UD object. The names of list will be used as layer names, usually they are model names.

hr_levels
The vector of level.UD in ctmm::plot.telemetry. To be consistent with ctmm they are values 0 ~ 1 (for example 0.95). Note the app UI take values 0 ~ 100 (for example 95) for easier input.

hr_color_vec
Vector of color names to be used for each home range. The length of hr_color_vec should match length of hrange_list.

Value

A Leaflet map widget.