Sample From Telemetry Object

A sampled dataset can have models fitted much quicker. This is used to reduce waiting time in developing code that involved time consuming modeling processes. After code is tested and stablized, full size dataset can be used.

sample_tele(tele, m)

Arguments

tele
telemetry object

m
sample size. m even spaced points are taken from data.

Value

sampled telemetry object

Examples

library(ctmm)
data(buffalo)
sample_tele(buffalo[[1]], 100)

#> Object of class "telemetry"
#>                timestamp longitude  latitude          t        x           y
#> 4109 2005-07-14 05:35:00  31.88776 -24.96738 1121319300 35215.76   836.73384
#> 4145 2005-07-15 17:34:00  31.87579 -25.03295 1121448840 42285.43 -1344.32858
#> 4181 2005-07-17 05:34:00  31.86597 -24.97956 1121578440 36262.64 -1524.77997
#> 4216 2005-07-18 16:34:00  31.90056 -24.98533 1121704440 37369.78  1846.83826
#> 4253 2005-07-20 04:34:00  31.87783 -24.97855 1121834040 36312.63  -323.89246
#> 4290 2005-07-21 16:34:00  31.90760 -24.98516 1121963640 37447.08  2553.36160
#> 4325 2005-07-23 03:33:00  31.91670 -25.00995 1122089580 40305.44  3090.12579
#> 4361 2005-07-24 15:33:00  31.88027 -24.99701 1122219180 38381.90  -357.07695
#> 4399 2005-07-26 02:33:00  31.88048 -24.98886 1122345180 37485.79  -213.58077
#> 4435 2005-07-27 14:33:00  31.87343 -24.99912 1122474780 38521.58 -1072.53710
#> 4471 2005-07-29 02:33:00  31.88417 -25.00311 1122604380 39107.98   -58.78622
#> 4507 2005-07-30 13:33:00  31.88185 -25.04201 1122730380 43367.16  -874.60867
#> 4543 2005-08-01 01:32:00  31.87819 -24.98110 1122859920 36598.71  -326.14957
#> 4580 2005-08-02 13:32:00  31.90555 -24.95472 1122989520 34061.58  2805.88672
#> 4615 2005-08-04 00:32:00  31.90828 -24.97235 1123115520 36043.26  2813.89808
#> 4652 2005-08-05 12:31:00  31.85851 -24.97305 1123245060 35443.32 -2173.15059
#> 4687 2005-08-06 23:31:00  31.88311 -24.99598 1123371060 38306.99   -57.68840
#> 4723 2005-08-08 11:32:00  31.87755 -24.98939 1123500720 37504.48  -514.65603
#> 4759 2005-08-09 23:31:00  31.87426 -24.99526 1123630260 38107.16  -931.63933
#> 4796 2005-08-11 10:31:00  31.86273 -25.06532 1123756260 45678.37 -3134.95495
#> 4832 2005-08-12 22:31:00  31.86904 -25.04510 1123885860 43533.89 -2201.05766
#> 4867 2005-08-14 09:30:00  31.86487 -24.99237 1124011800 37660.82 -1826.98033
#> 4903 2005-08-15 21:30:00  31.91367 -24.98852 1124141400 37900.28  3109.60415
#> 4939 2005-08-17 08:30:00  31.92389 -25.03401 1124267400 43057.23  3446.79056
#> 4974 2005-08-18 19:30:00  31.91697 -25.03316 1124393400 42869.12  2768.08509
#> 5010 2005-08-20 06:30:00  31.97952 -25.10893 1124519400 52079.68  7875.60692
#> 5046 2005-08-21 18:30:00  31.90431 -25.09074 1124649000 49047.79   637.33237
#> 5081 2005-08-23 05:29:00  31.91776 -25.02249 1124774940 41702.89  3007.48019
#> 5117 2005-08-24 17:29:00  31.91363 -25.02163 1124904540 41551.90  2607.60745
#> 5152 2005-08-26 03:29:00  31.91511 -24.99871 1125026940 39043.95  3100.34900
#> 5188 2005-08-27 15:29:00  31.90083 -24.99644 1125156540 38598.95  1706.92898
#> 5225 2005-08-29 03:28:00  31.89285 -24.96102 1125286080 34583.56  1441.26595
#> 5260 2005-08-30 14:29:00  31.86088 -24.97910 1125412140 36142.73 -2026.84997
#> 5298 2005-09-01 02:28:00  31.85546 -25.05038 1125541680 43931.74 -3637.32977
#> 5333 2005-09-02 13:28:00  31.85525 -25.06946 1125667680 46033.45 -3944.48899
#> 5370 2005-09-04 01:28:00  31.85525 -25.04256 1125797280 43066.21 -3541.23764
#> 5406 2005-09-05 13:28:00  31.86352 -25.07313 1125926880 46550.62 -3173.21786
#> 5441 2005-09-07 00:28:00  31.87253 -25.06409 1126052880 45675.99 -2137.31428
#> 5477 2005-09-08 12:27:00  31.86844 -25.08076 1126182420 47459.04 -2795.96348
#> 5513 2005-09-10 00:27:00  31.88527 -25.07106 1126312020 46617.81  -969.00029
#> 5550 2005-09-11 11:27:00  31.87064 -25.05752 1126438020 44925.44 -2227.57030
#> 5586 2005-09-12 23:27:00  31.83791 -25.03622 1126567620 42131.55 -5179.00667
#> 5621 2005-09-14 10:27:00  31.81696 -25.05254 1126693620 43647.53 -7517.08933
#> 5657 2005-09-15 22:26:00  31.83493 -25.03418 1126823160 41866.01 -5446.33590
#> 5693 2005-09-17 10:26:00  31.81899 -25.04612 1126952760 42966.98 -7218.15093
#> 5728 2005-09-18 21:26:00  31.84842 -25.05490 1127078760 44334.57 -4408.69211
#> 5764 2005-09-20 09:26:00  31.84249 -25.03947 1127208360 42552.10 -4769.97405
#> 5801 2005-09-21 20:26:00  31.82975 -25.04691 1127334360 43199.96 -6154.60308
#> 5837 2005-09-23 09:25:00  31.81518 -25.04025 1127467500 42267.87 -7511.17663
#> 5874 2005-09-24 21:25:00  31.82549 -25.02954 1127597100 41226.28 -6320.40627
#> 5909 2005-09-26 08:25:00  31.85265 -25.03380 1127723100 42064.58 -3669.57020
#> 5945 2005-09-27 20:25:00  31.84503 -25.04639 1127852700 43350.04 -4619.91946
#> 5982 2005-09-29 08:24:00  31.84463 -25.05039 1127982240 43785.83 -4719.69468
#> 6017 2005-09-30 19:25:00  31.85348 -25.06363 1128108300 45366.24 -4033.93916
#> 6053 2005-10-02 07:24:00  31.86689 -25.05580 1128237840 44684.86 -2576.45286
#> 6088 2005-10-03 18:24:00  31.92207 -25.02739 1128363840 42302.23  3364.62208
#> 6124 2005-10-05 06:24:00  31.91348 -25.01155 1128493440 40437.96  2744.35895
#> 6160 2005-10-06 18:24:00  31.98389 -25.12322 1128623040 53715.51  8096.33365
#> 6195 2005-10-08 05:24:00  31.91292 -25.10916 1128749040 51196.55  1220.43732
#> 6231 2005-10-09 17:23:00  31.91432 -24.99209 1128878580 38302.98  3120.99578
#> 6266 2005-10-11 04:23:00  31.89861 -24.95388 1129004580 33874.34  2124.52356
#> 6302 2005-10-12 15:23:00  31.92132 -25.02802 1129130580 42361.34  3280.13492
#> 6338 2005-10-14 03:23:00  31.86099 -25.01148 1129260180 39716.02 -2501.42833
#> 6373 2005-10-15 14:23:00  31.83318 -25.05343 1129386180 43965.76 -5909.43906
#> 6409 2005-10-17 01:22:00  31.83640 -25.04321 1129512120 42882.15 -5434.77063
#> 6445 2005-10-18 13:22:00  31.82402 -25.04505 1129641720 42917.12 -6699.52456
#> 6480 2005-10-20 00:22:00  31.82142 -25.04958 1129767720 43381.66 -7027.16752
#> 6516 2005-10-21 12:22:00  31.91451 -25.02644 1129897320 42094.40  2623.32555
#> 6551 2005-10-22 23:22:00  31.95825 -25.12658 1130023320 53736.03  5485.19376
#> 6587 2005-10-24 11:21:00  31.94623 -25.13377 1130152860 54365.05  4176.59755
#> 6623 2005-10-25 23:21:00  31.94792 -25.15482 1130282460 56709.72  4028.29242
#> 6659 2005-10-27 10:21:00  31.94567 -25.14684 1130408460 55798.95  3923.84702
#> 6695 2005-10-28 22:21:00  31.90953 -24.98317 1130538060 37253.89  2776.10397
#> 6731 2005-10-30 09:21:00  31.94987 -25.00931 1130664060 40687.27  6415.34038
#> 6767 2005-10-31 21:20:00  31.93668 -25.00549 1130793600 40085.82  5154.50072
#> 6803 2005-11-02 09:20:00  31.91849 -25.02352 1130923200 41826.61  3064.96333
#> 6838 2005-11-03 19:20:00  31.91107 -25.03144 1131045600 42598.88  2204.32389
#> 6874 2005-11-05 07:20:00  31.90719 -25.00644 1131175200 39788.54  2192.39518
#> 6910 2005-11-06 19:20:00  31.87751 -24.94703 1131304800 32831.46   117.25012
#> 6945 2005-11-08 06:19:00  31.87963 -24.90326 1131430740 28032.32   986.27414
#> 6981 2005-11-09 18:19:00  31.89786 -24.91208 1131560340 29253.66  2677.47784
#> 7017 2005-11-11 05:19:00  31.91518 -24.92831 1131686340 31279.85  4166.01558
#> 7053 2005-11-12 17:19:00  31.88469 -24.97398 1131815940 35901.95   430.65114
#> 7089 2005-11-14 05:19:00  31.88092 -24.95593 1131945540 33859.67   324.65573
#> 7124 2005-11-15 16:18:00  31.86290 -24.98809 1132071480 37161.88 -1959.67774
#> 7160 2005-11-17 04:18:00  31.86110 -24.96149 1132201080 34203.39 -1740.83254
#> 7195 2005-11-18 15:18:00  31.85789 -24.99633 1132327080 38002.66 -2584.21175
#> 7231 2005-11-20 03:18:00  31.84774 -24.95592 1132456680 33407.37 -2993.44735
#> 7267 2005-11-21 15:18:00  31.84779 -24.94269 1132586280 31948.66 -2790.10863
#> 7302 2005-11-23 02:18:00  31.87122 -25.00960 1132712280 39647.71 -1450.68080
#> 7338 2005-11-24 14:17:00  31.89088 -24.98431 1132841820 37125.53   894.33725
#> 7374 2005-11-26 02:17:00  31.88891 -24.97054 1132971420 35579.91   904.22563
#> 7409 2005-11-27 13:17:00  31.89533 -25.02029 1133097420 41154.80   798.86280
#> 7445 2005-11-29 00:17:00  31.92897 -25.05131 1133223420 45034.58  3694.11838
#> 7482 2005-11-30 11:17:00  31.93402 -25.05101 1133349420 45070.25  4203.42274
#> 7518 2005-12-01 23:17:00  31.94870 -25.06084 1133479020 46354.81  5522.03999
#> 7555 2005-12-03 11:16:00  31.96414 -25.06163 1133608560 46652.68  7052.75250
#> 7591 2005-12-04 22:16:00  31.95774 -25.05224 1133734560 45529.69  6554.80389
#> 7627 2005-12-06 10:16:00  31.94410 -25.08189 1133864160 48613.88  4745.35526
#> 7663 2005-12-07 22:16:00  31.90330 -25.04151 1133993760 43603.87  1276.45307
#> Slot "info":
#> $identity
#> [1] "Cilla"
#> 
#> $timezone
#> [1] "UTC"
#> 
#> $projection
#> [1] "+proj=tpeqd +lon_1=31.7775407470691 +lat_1=-24.2850983757726 +lon_2=31.8880063438124 +lat_2=-25.024736901024 +datum=WGS84"
#> 
#> $UERE
#> NULL
#> 
#>