Generate formated model summary table from tried models results

Generate formated model summary table from tried models results

summary_tried_models(model_try_res)

Arguments

model_try_res
list of applying ctmm::ctmm.select on telemetry objects

Value

A data.table of model summary