Generate Merged Location And Info data.table From Telemetry Object/List

A Telemetry list hold mutiple animal data in separate list items, each item have the animal location data in a data frame, and other information in various slots. This structure supports flexible S3 methods for telemetry object. However to plot multiple animals location together with ggplot2 it's better to merge all location data into single data frame with an animal id column.

merge_animals(tele_obj_list)

Arguments

tele_obj_list
telemetry object/list

Value

list of

data: all animals merged in one data.table

info: animal information table

Details

We chose this data structure to be the main structure and made the app to work on a set of animals at the same time in all steps. Thus any input telemetry object/List need to be merged into a data.table of location data, and another information data.table for animals. data.table is chosen over data.frame for much better performance.

Examples

library(ctmm)
data(buffalo)
merge_animals(buffalo)

#> $data
#>                  timestamp longitude  latitude          t         x          y
#>     1: 2005-07-14 05:35:00  31.88776 -24.96738 1121319300  35215.76   836.7338
#>     2: 2005-07-14 07:35:00  31.85942 -24.94288 1121326500  32127.59 -1629.8224
#>     3: 2005-07-14 08:34:00  31.85512 -24.94914 1121330040  32759.68 -2153.6281
#>     4: 2005-07-14 09:35:00  31.85694 -24.95424 1121333700  33347.02 -2048.1883
#>     5: 2005-07-14 10:35:00  31.85977 -24.94735 1121337300  32625.39 -1661.8450
#>    ---                                                                        
#> 17338: 2006-04-22 19:09:00  31.74776 -24.20799 1145732940 -50456.95 -1845.2713
#> 17339: 2006-04-22 20:10:00  31.74706 -24.20839 1145736600 -50422.24 -1921.6635
#> 17340: 2006-04-22 21:10:00  31.74524 -24.20557 1145740200 -50758.07 -2062.8504
#> 17341: 2006-04-22 22:10:00  31.74522 -24.20555 1145743800 -50760.56 -2064.3894
#> 17342: 2006-04-22 23:09:00  31.74524 -24.20554 1145747340 -50761.44 -2062.3955
#>        identity row_name    id row_no
#>     1:    Cilla     4109 Cilla      1
#>     2:    Cilla     4110 Cilla      2
#>     3:    Cilla     4111 Cilla      3
#>     4:    Cilla     4112 Cilla      4
#>     5:    Cilla     4113 Cilla      5
#>    ---                               
#> 17338:     Toni    17512  Toni  17338
#> 17339:     Toni    17513  Toni  17339
#> 17340:     Toni    17514  Toni  17340
#> 17341:     Toni    17515  Toni  17341
#> 17342:     Toni    17516  Toni  17342
#> 
#> $info
#>    identity interval_s   interval duration_s   duration      sampling_start
#> 1:    Cilla       3600 60 minutes   12674460   5 months 2005-07-14 05:35:00
#> 2:     Gabs       3600 60 minutes    7163340 2.8 months 2005-04-05 05:56:00
#> 3:    Mvubu       3600 60 minutes    9208020 3.6 months 2005-07-15 05:02:00
#> 4:   Pepper       7200    2 hours   21633900 8.5 months 2006-04-25 05:09:00
#> 5:    Queen       3600 60 minutes    9067080 3.6 months 2005-02-17 05:05:00
#> 6:     Toni       3600 60 minutes   20968440 8.2 months 2005-08-23 06:35:00
#>           sampling_end            start              end points
#> 1: 2005-12-07 22:16:00 2005-07-14 05:35 2005-12-07 22:16   3527
#> 2: 2005-06-27 03:45:00 2005-04-05 05:56 2005-06-27 03:45   1996
#> 3: 2005-10-29 18:49:00 2005-07-15 05:02 2005-10-29 18:49   2572
#> 4: 2006-12-31 14:34:00 2006-04-25 05:09 2006-12-31 14:34   1725
#> 5: 2005-06-02 03:43:00 2005-02-17 05:05 2005-06-02 03:43   1756
#> 6: 2006-04-22 23:09:00 2005-08-23 06:35 2006-04-22 23:09   5766
#>